fetch и разбивка по view
