+++ public/js/views/Menu.js
@@ -1,6 +1,6 @@
+import * as ViewsContext from "./ViewsContext.js";
 import {MenuComponent} from "../components/Menu/Menu.mjs";
 import {SectionComponent} from "../components/Section/Section.mjs";
-import * as ViewsContext from "./ViewsContext.js"
 
 export const showMenu = () => {
     const content = document.querySelector('.content');
@@ -9,11 +9,11 @@ export const showMenu = () => {
     menuSection.render();
 
     const titles = new Map();
-    titles.set('index', 'Играть');
+    titles.set('play', 'Играть');
     titles.set('profile', 'Профиль');
     titles.set('scoreboard', 'Списки лидеров');
     titles.set('about', 'Об игре');
-    titles.set('login', 'Выход');
+    titles.set('logout', 'Выход');
 
     const menu = new MenuComponent({
         el: menuSection.sectionContent,
@@ -21,7 +21,7 @@ export const showMenu = () => {
         actionOnButton: (event) => {
             event.preventDefault();
             const link = event.target;
-            hideAnySection();
+            ViewsContext.hideAnySection();
             ViewsContext.pages[ link.dataset.href ]();
         }
     });
